+++ CHANGELOG.md
@@ -3,6 +3,8 @@ GAMSPy CHANGELOG
 
 GAMSPy 1.0.0rc1
 ---------------
+- General
+  - Include variable infeasibilities in model.computeInfeasibilities().
 - Documentation
   - Add gamspy probe and gamspy retrieve to the cli reference page.